+++ sdk/databricks/arm-databricks/src/classic/index.ts
@@ -3,7 +3,6 @@
 
 export type { AccessConnectorsOperations } from "./accessConnectors/index.js";
 export type { OperationsOperations } from "./operations/index.js";
-export type { OutboundNetworkDependenciesEndpointsOperations } from "./outboundNetworkDependenciesEndpoints/index.js";
 export type { PrivateEndpointConnectionsOperations } from "./privateEndpointConnections/index.js";
 export type { PrivateLinkResourcesOperations } from "./privateLinkResources/index.js";
 export type { VNetPeeringOperations } from "./vNetPeering/index.js";